refactor: 메뉴 상품 목록 필드를 일급 컬렉션으로 변경
메뉴 상품 도메인과 상품의 요구사항을 처리하는 중간 계층 정의
- MenuProductService.java
- 메뉴 상품 요청 객체로부터 상품을 조회하여 메뉴 상품 객체를 산출하는 기능 구현

+++ src/main/java/kitchenpos/dto/menu/MenuProductRequest.java
@@ -7,7 +7,7 @@ import kitchenpos.domain.product.Product;
 public class MenuProductRequest {
 
     private Long productId;
-    private int quantity;
+    private long quantity;
 
     public MenuProductRequest() {
 
@@ -22,7 +22,7 @@ public class MenuProductRequest {
         return productId;
     }
 
-    public int getQuantity() {
+    public long getQuantity() {
         return quantity;
     }
